requirements.txt
openai
langchain
langchain_text_splitters
pymilvus
streamlit
python-dotenv
numpy